Add files via upload
This the GreenElephant Library for C#. It is mostly a math library that adds a lot of functionality in C#. More specifically: Complex number, easy conversion between coordinate systems, converting a string to double-float and more. The library is in working progress.

To use it, download the file and put into your project. Then write at the start of your code "using GreenElephant;".

Feel free to ask any questions, if you have got any problems.